Easy OSTF launching based on file config
The presented patch provides several necessary lines of code that make it
possible to run the OSTF tests based on a config written in a file. All the
user has to do is to insert an instruction like

os.environ['CUSTOM_FUEL_CONFIG'] = 'path/to/config/file'

before the actual OSFT launch. Changing the FuelConfig() method
implementation as well as moving its call in the BaseTestCase class
into the setUpClass() method serves for this goal and ensures that
the config values will override the default ones.