Fix test row execution
Logic that controlls test execution has been modified. Now all tests
have priority in exectution that controlls execution type. So now there
are excluseve tests which are started in child subprocesses and deffered
tests that are executed in same processes after exclusive finished.
Addition refactoring was added for testruns controller and TestRun
entity logic so that now when there are no exclusive test_sets in start
request body all of available are considered exlusive and are executed
simulateously.

Change-Id: If6d61515e9c3847366102d53534634c96f4d204e