4.1.1 Fixed issue with parrallel tests execution
When user runs several types of OSTF tests, these tests can be run
in parallel and in the result tearDown method deletes all environment
and some tests will fail. To avoid this we will generate new name
for each new environment and we should also delete only this environment
in tearDown method.

Change-Id: I12f1e797eceefe11dda50ab61aabe15ab9036931
Closes-Bug: #1285562