Row tests exectuion via locking mechanism added
Now testruns are created at the same time but first of dependent aquire
lock on file descriptor. This allows to separate execution of dependent
testruns. Lock is corresponding to particular set of tests.

Each of dependent testsets store names of series that it belongs to on
exclusive_testsets attribute. In this case testrun have to aquire
all locks which are corresponding to those series.

Change-Id: Ia248058dbec336ba2fad7703cc6a1aaf67d8e36e