Fix for SSH module of Sahara OSTF tests
In SSH method of Sahara OSTF tests address of controller node is chosen
as the first from list of controller nodes.
It may not work when we have a few controller nodes (HA mode) and one
of controller nodes went away. Now controller node is chosen
from list of controller nodes online.

Closes-Bug: #1346864

Change-Id: I74801a352c4a96e1fa1c254fd49db19d6584b427