5.0 Fix for SSH method of OSTF tests
In SSH method of OSTF tests address of controller node is chosen
as the first from list of controller nodes.
It may not work when we have a few controller nodes (HA mode) and one
of controller nodes went away. Now controller node is chosen
from list of controller nodes online.

Change-Id: I7f9eedfc797689e609680016a7085fdf30af24e7