Populate test cache if it's empty
In commit e232d498a4c5fd8fbf4ca9c17f528ac6274146b3 bug (#1367359)
when there were no TestRuns for cluster when OSTF tests were started
without asking for a lost of available TestRuns.

In the same case when tests cache is empty not all required
objects (ClusterTestingPattern) will be created in DB that are
needed to correctly run tests.

Now we force cache repopulation if it's empty.

Additionally removed some dead code that is a duplicate
of modified one in this changestet.

Change-Id: If7dd352f2a48c27b7aa9f41c6c5a4528021a8208
Closes-Bug: #1387733