Populate test cache if it's empty
In some rare cases test cache in OSTF might be empty and
not all required objects will be created in DB that are
needed to correctly run tests.
Now we force cache repopulation if it's empty.

Additionally removed also some dead code that is a duplicate
of modified in this changestet.

Change-Id: If7dd352f2a48c27b7aa9f41c6c5a4528021a8208
Closes-Bug: #138733