Increase timeout for snapshotting from 180 to 700 s
1. [compute/build_timeout] stanza in OSTF config file is used not for build
VM only but for VM snapshot creation too.
Nova-compute spend a significant time to download/upload big images
to/from vCenter (5-10MB/s in our testbed). So it leads to a timeout error
for 'boot_from_snapshot' operation with the 180s timeout.
That's why the [compute/build_timeout] is increased to 500s.
It shouldn't affect any other test because the 'status_timeout' method
(from test.py) which uses this stanza always wrapped with 'verify' method
(from test_mixins.py), and the 'verify' method is always called with more
strict (and unique) timeout for every step in each test.

2. Timeouts for snapshot operations should be increased in case of
   vCenter as a hypervisor.for They are fixed from 180s to 700s.

3. Logging improvements in 'verify' method (from test_mixins.py).

Change-Id: I35521f75dae4f9bef4e20fd8c1154f2e510ba921
Closes-bug: 1386547