fix check_inet_conn_instance_without_floatingIP
There is a parameter compute in this test, which
calculated as
compute = getattr(server, 'OS-EXT-SRV-ATTR:host')

After http://bit.ly/1ymKAYH on vmware based envs
meaning of the parameter host was changed and now it
just index [0, 1, 2, ...] and calculate take such
number as a value. As a result ssh obligatory
fails when it trying  connect to host
"ssh root@0".

See http://bit.ly/1J8HbOR for detail.

Calculate isn't mandatory parameter, it has a
default value None with which test should pass.

The patch removes this parameter.

Change-Id: Ie1068c38324d24e2a9d511992d1b56555748f590
Closes-bug: #1405493