fix check_inet_conn_instance_without_floatingIP
There is a parameter compute in this test, which
calculated as
compute = getattr(server, 'OS-EXT-SRV-ATTR:host')

After http://bit.ly/1ymKAYH on vmware based
environments meaning of the parameter host was
changed and now it just index [0, 1, 2, ...] and
calculate take such number as a value. As a
result ssh obligatory fails when it trying to
connect to host "ssh root@0".

See http://bit.ly/1J8HbOR for detail.

There are no nodes with role 'compute' on vmware
based environment and in this case calculate
isn't mandatory parameter. When we set it to None
test should pass.

Change-Id: Ie1068c38324d24e2a9d511992d1b56555748f590
Closes-bug: #1405493