Add versioning for fuel_health tests
Now version of release for which particular test is suitable can be
specified in test sets profile attribute and tests' docstrings. For now
we assume that test is available for releases that are == or >= in
comparison to release specified for test.

Now cleaning of db is before executing of integration testsing order to prevent issue with reapplying head migration after revision script for it was changed.

Change-Id: I568db380120e3a429bdd531f51a7301c49d41281
Implements: blueprint ostf-tests-versioning