Fixing platform test for Sahara
If we create a private network without subnet and DHCP agent,
this network may be chosen by the test as the private network
for Sahara cluster. In this case the test will fail. Now the
private network for cluster will be chosen by the test if this
network has a subnet and an enabled DHCP agent.

Change-Id: I8fdac844eea1f48b3d6ea8e63c8adf11391b2cc9
Closes-Bug: #1421739