Fixing platform test for Sahara
If we create a private network without subnets or for example,
DHCP agents, this network may be chosen by the test as the private
network for Sahara cluster. In this case the test will fail. Now
the default private network is chosen as the private network for
Sahara cluster. This fix also affects Ceilometer tests.
So Ceilometer tests have the corresponding changes.

Closes-Bug: #1421739

Change-Id: I8fdac844eea1f48b3d6ea8e63c8adf11391b2cc9