Editing link to docs for Sahara OSTF
For now platform test for Sahara throws an exception with the link
to section "Platform Tests Description" if the test doesn't find correct
image for Sahara. It is not very convinient for users. Now platform test
for Sahara throws the exception with the link to section
"Preparing Sahara for Testing".

Change-Id: I3e2f2e2e737c7a1e48b1dbd835b961db2c00783c
Closes-Bug: #1337861