Fixed order of checks in the Heat OSTF test
In Heat OSTF test 'Advanced stack actions' there was wrong order
of checks during step 9. We firstly checked status of resource
and only after that status of stack, that's why first check failed
(because the resource does not have time to go to the desired status).
Now we change order os checks and firstly wait until stack status will
change to 'CHECK_COMPLETE' - it is a guarantee that all stack resources
also have to be in 'CHECK COMPLETE' status.

Partially-Implements: blueprint heat-ostf-test-refactoring

Change-Id: I524de4fc4a373558a098fcb6e37d12c143fa0416