Use service_endpoint as an ednpoint for working with openstack cluster
To enable granular deployment of keystone, we need to know
precisely where keystone will be configured, as we dont have
such API in nailgun, we need to invent some convention, which
would be <service_name>_vip to access granular service.

Important to remember that we also need proxy on port 8888 configured
on a node that consumes configured vip.

Related-Bug: #1470927
Related bp ostf-separate-services-refactor
Change-Id: Ie497bca8818cd0f24eb40df32df227152254a69d