Use ssl settings for horizon tests if enabled
This change uses the ssl configuration from the environment for the
Horizon testing if ssl is enabled. If ssl is enabled for horizon, the
test will connect to the ssl version of Horizon but will not validate
the SSL certificate that is in place.

DocImpact: OSTF Horizon test leverages the Public TLS settings for the
environment to test either the HTTP or HTTPS endpoint as part of the
"Create user and authenticate with Horizon" test.

Change-Id: Iee86c6ef0c4dc9d7caf7148b1358972561b25a96
Closes-Bug: 1486056