Use ssl settings for horizon tests
This change uses the ssl configration from the environment for the
Horizon testing. This test will validate the SSL certificate and uses
the hostname from the settings to connect to Horizon.

DocImpact: OSTF Horizon uses the hostname provided in the SSL
configuration section for the environment to test with. If this name
does not match the certificate or cannot be resolved from the master,
this test will fail.
Closes-Bug: 1486056

Change-Id: Iee86c6ef0c4dc9d7caf7148b1358972561b25a96