Use ssl settings for horizon tests if enabled
This change uses the ssl configuration from the environment for the
Horizon testing if ssl is enabled. If ssl is enabled for horizon, the
test will validate the ssl certificate and use teh hostname provided in
the Public TLS settings for the environment.

DocImpact: OSTF Horizon uses the hostname provided in the SSL
configuration section for the environment to test with when SSL is
enabled. If this name does not match the certificate or cannot be
resolved from the master, this test will fail. For Horizon without
SSL, the IP address is used for the test.

Change-Id: Iee86c6ef0c4dc9d7caf7148b1358972561b25a96
Closes-Bug: 1486056