Fix problem with race condition of forked ostf
When there is a more then one check in ostf it fails
and only one process works. Other forks are dead because
they can not fetch information from database.
Eventually information is flushed in database, but only
after forks are dead.
Sleep helped to wait for information to propagate to
other processes. This fix flush and commit information
to database so it is available for other forks.

Change-Id: Ib8540309dee00f7f0ea7cd65215c4872c4147b6f
Closes-bug: #1520218