Adjust heat test timeouts
This change increases the timeout for the heat create stack action from
20 seconds to 60 seconds. The UCA provided heat seems to take longer
than 20 seconds for the stack to be created and when the timeout hits,
all of the associated networks and flavors disappear which leads to
stack failures. Additionally this change increases the timeout for the
wait for the stack to be deleted from 180 to 300 as it appears the
creation and deletion from a rollback takes longer under UCA packages.

Change-Id: I0f0fe597d125ef71764fc239055de59e33a92b75
Related-Blueprint: deploy-with-uca-packages