Add processing of ssl certificate
Now there is possibility to enable ssl for nginx by adding 'flat_https'
option to fuel settings (/etc/fuel/astute.yaml on the master node).
Since ostf adapter performs http requests to nailgun in order to obtain
particular info from it, ssl cerificate that is used by nginx should be
supplied to the corresponding adapter's code for verification. But this
must be done only when ssl is actually enabled so code that check it
added too.

Change-Id: I98733ba57e87c3b59aeddc4f1e601a5518aeb439
Closes-Bug: #1530318