Fix ironic check_service_availability
run_ssh_cmd_with_exit_code mark test as failed as soon
it executes command on node where process is not running.
As result when process is not running on first node
test is failed, and other nodes are not checked.

check_service_availability should suceed only if process
is running on specified number of nodes.

ironic-api should be launched on all controllers.
ironic-conductor should be launched on all Ironic nodes.
nova-compute should be launched only on single controller.

Change-Id: I5429f079d903e51cbe786548126af07bc140aee0