Fix ironic check_service_availability
check_service_availability should fail only if process
is not running on all nodes.
run_ssh_cmd_with_exit_code mark test as failed as soon
it executes command on node where process is not running.
As result when process is not running on first node
test is failed, and other nodes are not checked.

Change-Id: I5429f079d903e51cbe786548126af07bc140aee0