Run the file injection test case on Ceph envs
Recently we changed the way how file injection is performed: instead
of using nbd or libguestfs, which both modify the image in-place
(and, thus, do not work with Ceph ephemerals), we now place files to
be injected on a config drive and force instances to have config
drives in Nova.

This test case should pass now, when Ceph ephemerals are used. Let's
enable it and actually test, that file injection works as expected.

Related-Bug: #1467860
Related-Bug: #1556819

Change-Id: I5975ea6d481c28856efffb278112306c8760472d