Removed self.doc_link from fuel platform tests for murano (linux)
The URL for self.doc_link was outdated, and had been moved.
causing the link to 404 when clicked.

self.doc_link was removed in order to keep consistency across
the project (other instances had been removed) and URLs are hard
to maintain across releases.

The message now says: "Murano image was not properly regstered or
was not uploaded at all. Please refer to the Mirantis Openstack
documentation to find out how to upload and/or how to register
image for Murano"

Change-Id: I124a2b657a39122d1801136ccda3f5d3bda9a2c4
Closed-bug: #1588211